PO classes have been moved to the main folder. Package names are given to the naming standard. Modified file README.md
