+++ src/main/java/ru/praktikum/service/qascooter/po/HeadPage.java
@@ -1,4 +1,4 @@
-package ru.praktikumService.qaScooter;
+package ru.praktikum.service.qascooter.po;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
